Fix IAP for iOS: apply()'s "this" argument needs to be set.
When I changed the JS callbacks to use `apply()`, I left the
"this" argument set to null. This caused code in InAppPurchaseManager.prototype.updatedTransactionCallback
to fail because references to `this` become null instead
of references to the default instance of InAppPurchaseManager.

This commit modifies the apply() calls to use plugins.inAppPurchaseManager as the "this" argument.